fix : 테이블뷰 수정
