+++ src/pages/MainPage.js
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { Heart, Star, Calendar, TrendingUp, Zap, Film, ArrowUp } from 'lucide-react';
+import { Heart, Star, Calendar, TrendingUp, Zap, Film } from 'lucide-react';
 import Header from '../components/common/header';
 import MovieSection from '../components/main/MovieSection';
 import ScrollToTopButton from '../components/common/ScrollToTopButton';
